Create a Custom Keymap for the Corne V4
This is an attempt at a custom keymap for the Corne V4 I got from
AliExpress. This firmware is untested. I discovered that the board
contains a Geehy APM32F103xx which does not match the board expected
by QMK (RP2020). Thus, it looks like VIA is the only way to control
the firmware, which is undesirable. Thus, I am attempting to return
the keyboard. Maybe in the future I will come back to this.